requirements.txt
# Core dependencies for RAG pipeline
numpy>=1.24.0
PyYAML>=6.0
tqdm>=4.65.0

# Tokenization and chunking
chonkie>=1.0.2,<1.1.0
tiktoken>=0.5.0

# Embeddings and models
sentence-transformers>=2.2.0
transformers>=4.40.0
torch>=2.0.0

# Vector search
faiss-cpu>=1.8.0  # or faiss-gpu if GPU available

# Sparse retrieval
bm25s[core]==0.2.1

# Reranking
# cross-encoder models are loaded via sentence-transformers/transformers

# Evaluation
scikit-learn>=1.3.0

# Optional: for advanced BM25
# pyserini>=0.22.0

# For semantic chunking and table processing
# (no additional dependencies - uses existing libraries)

# Optional: for debugging panel
# streamlit>=1.28.0
# gradio>=4.0.0

